% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{exercise}
\alias{exercise}
\title{Get an exercise}
\usage{
exercise(name = c("faux", "fixed", "mixed", "calories"), filename = NULL)
}
\arguments{
\item{name}{The name of the exercise}

\item{filename}{What filename you want to save (defaults to the name of the exercise in the working directory)}
}
\value{
Saves a file to the working directory (or path from filename)
}
\description{
Get an exercise
}
\examples{
\dontrun{
exercise("faux") # get exercise for the faux workshop
exercise("fixed", "exercises/fixed.Rmd") # save into exercises directory
}
}
